+++ 4.Development/web/ej_manager/src/pages/product/Product.vue
@@ -3,16 +3,10 @@
         <!-- 按钮 -->
         <el-button type="primary" size="small" @click='toAddHandler'>添加</el-button>
         <div class="select">
-            <el-select v-model="categoryForm" placeholder="请选择">
-            <el-option
-            v-for="item in categorys"
-            :key="item.id"
-            :value="item.id"
-            :label="item.name"
-            >
-            </el-option>
-        </el-select>
-        <el-button type="primary" size="small">查询</el-button>
+            <el-input  v-model="list.name" clearable=true placeholder='请输入产品名称'></el-input>
+            <div class='btn'>
+            <el-button type="primary" @click="submitHanlderbyStatus" size="small">查询</el-button>
+            </div>
         </div>
         <!-- 按钮 -->
         <!-- 表格 -->
@@ -62,8 +56,10 @@
         <!-- 表格 -->
         <!-- 分页 -->
         <el-pagination
-            layout="prev, pager, next"
-            :total="50">
+            layout="total,prev, pager, next"
+            @current-change="changePageNum"
+            :page-size="this.list.pageSize"
+            :total="total">
         </el-pagination>
         <!-- 分页 -->
         <!-- 模态框 -->
@@ -112,6 +108,10 @@ import {mapState,mapActions} from 'vuex'
 export default {
     data(){
         return {
+            list:{
+                page:0,
+                pageSize:6
+            },
             title:'添加产品信息',
             dialogFormVisible:false,
             form:{},
@@ -135,15 +135,15 @@ export default {
         }
     },
     created() {
-        this.findAllproduct(),
+        this.queryProduct(this.list),
         this.findAllcategory()
     },
     computed: {
-        ...mapState('product',['products']),
+        ...mapState('product',['products','total']),
         ...mapState('category',['categorys']),
     },
     methods: {
-        ...mapActions('product',['findAllproduct','saveProduct','deleteProduct']),
+        ...mapActions('product',['findAllproduct','saveProduct','deleteProduct','queryProduct']),
         ...mapActions('category',['findAllcategory']),
         //添加打开模态框
         toAddHandler(){
@@ -173,12 +173,22 @@ export default {
         },
         //删除订单信息
         deleteHandler(id){
+            this.$confirm('此操作将永久删除这条数据, 是否继续?', '提示', {
+          confirmButtonText: '确定',
+          cancelButtonText: '取消',
+          type: 'warning'
+        }).then(() => {
             this.deleteProduct(id)
-            this.$message({
-            showClose: true,
-            message: '删除成功',
-            type: 'success'
-            });
+          this.$message({
+            type: 'success',
+            message: '删除成功!'
+          });
+        }).catch(() => {
+          this.$message({
+            type: 'info',
+            message: '已取消删除'
+          });          
+        });
         },
          //修改订单信息
         editHandler(custermerForm){
@@ -189,14 +199,29 @@ export default {
         //详情页面
         detailsHandler(id){
             this.$router.push({name:'productDetails',params:{id:id}})
+        },
+        //分页
+        changePageNum(page){
+                this.list.page = page-1;
+                this.queryProduct(this.list)
+        },
+        //模糊查询
+        submitHanlderbyStatus(){
+            this.queryProduct(this.list)
         }
     },
 }
 </script>
 <style scoped>
 .select{
+    width:200px;
     margin-left: 5em;
     margin-top: -2em;
+    margin-bottom:1em;
+}
+.btn{
+    margin-left: 13em;
+    margin-top: -34px;
 }
 .upload{
     margin-left:9em;
